Move Deferred to services-core
