+++ server/routerlicious/packages/lambdas-driver/src/kafka-service/checkpointManager.ts
@@ -4,8 +4,7 @@
  */
 
 import assert from "assert";
-import { Deferred } from "@fluidframework/server-common-utils";
-import { IConsumer, IQueuedMessage } from "@fluidframework/server-services-core";
+import { Deferred, IConsumer, IQueuedMessage } from "@fluidframework/server-services-core";
 import { Lumberjack } from "@fluidframework/server-services-telemetry";
 
 export class CheckpointManager {
