+++ server/routerlicious/packages/routerlicious-base/src/test/nexus/io.spec.ts
@@ -19,7 +19,6 @@ import {
 	IDocumentMessage,
 	type ISentSignalMessage,
 } from "@fluidframework/protocol-definitions";
-import { Deferred } from "@fluidframework/server-common-utils";
 import { KafkaOrdererFactory } from "@fluidframework/server-kafka-orderer";
 import { LocalWebSocket, LocalWebSocketServer } from "@fluidframework/server-local-server";
 import { configureWebSocketServices } from "@fluidframework/server-lambdas";
@@ -30,6 +29,7 @@ import {
 	clientConnectivityStorageId,
 	DefaultMetricClient,
 	DefaultServiceConfiguration,
+	Deferred,
 	IClientManager,
 	IOrdererManager,
 	MongoDatabaseManager,
