+++ server/routerlicious/packages/services-core/src/index.ts
@@ -33,6 +33,7 @@ export {
 	IRetryable,
 	isRetryEnabled,
 } from "./database";
+export { Deferred } from "./deferred";
 export { IDeltaService } from "./delta";
 export {
 	IClientSequenceNumber,
