+++ server/routerlicious/packages/services-core/src/queue.ts
@@ -3,7 +3,7 @@
  * Licensed under the MIT License.
  */
 
-import { Deferred } from "@fluidframework/server-common-utils";
+import { Deferred } from "./deferred";
 
 import { ITicketedMessage } from "./messages";
